Missing tests added.
