+++ IfWithoutBracketsAnalyzer/IfWithoutBracketsAnalyzer.Test/Resources/CodeFile.txt
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Diagnostics;
+
+namespace ConsoleApplication1
+{
+    class Class1
+    {
+        public void Foo()
+        {
+            var tmp = true;
+            if (tmp)
+                tmp = false;
+        }
+    }
+}
